refactor(历史备注): 收拢模型并适配弹窗缩放
1、将历史备注范围、排序和聚合结果收拢到models，解除偏好Provider对repository/service的反向依赖
2、仓储、同步和备注组件改为直接依赖历史备注领域模型，保持现有查询与同步语义
3、历史备注设置弹窗的字体、间距、输入宽度和内边距统一使用scaled适配显示缩放
