+++ lib/data/repositories/local/local_repository.dart
@@ -6,13 +6,12 @@ import '../../../cloud/sync/change_tracker.dart';
 import '../../../services/currency/rate_math.dart';
 import '../../../utils/shared_ledger_picker_filter.dart';
 import '../../../services/system/logger_service.dart';
+import '../../../models/note_history.dart';
 import '../base_repository.dart';
 import '../budget_repository.dart';
 import '../transaction_repository.dart'
     show
         BatchAttachmentData,
-        NoteHistoryEntry,
-        NoteHistorySort,
         TransactionUpdateBySyncIdData;
 import 'local_ledger_repository.dart';
 import 'local_transaction_repository.dart';
