+++ lib/providers/sync_providers.dart
@@ -21,8 +21,7 @@ import 'voice_billing_providers.dart';
 import '../services/attachment_service.dart' show attachmentListRefreshProvider;
 import '../services/system/logger_service.dart';
 import '../services/ui/avatar_service.dart';
-import '../data/repositories/transaction_repository.dart';
-import '../services/data/note_history_service.dart';
+import '../models/note_history.dart';
 import 'theme_providers.dart';
 import 'budget_providers.dart';
 import 'calendar_providers.dart';
