+++ lib/providers/theme_providers.dart
@@ -3,8 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import '../data/repositories/transaction_repository.dart';
-import '../services/data/note_history_service.dart';
+import '../models/note_history.dart';
 import '../services/system/logger_service.dart';
 import '../theme.dart';
 import '../widget/widget_manager.dart';
